chore: refactor to helper function + hook
